% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchability.R
\name{catchability}
\alias{catchability}
\title{This function calculates the probability of entry to trap (aka catchability). In addition to parameters in Addison & Bell
1997, length of catch can is added and can be included in calculation of catchability.}
\usage{
catchability(
  q0,
  qmin,
  saturationThreshold,
  Ct,
  lengthBased,
  lobLengthThreshold,
  lobSize = NA,
  sexBased,
  lobSex
)
}
\arguments{
\item{q0}{is the initial probability of entry into an empty trap(i.e. 0.5).}

\item{qmin}{is the asymptotic minimum probability of entry (i.e. 0).}

\item{saturationThreshold}{is the number of lobsters in a trap at which the probability of
another lobster entering the trap drops to qmin}

\item{Ct}{is the number of caught lobster}

\item{lengthBased}{is a logical parameter(TRUE/FALSE) which determines if length of catch needs too be considered}

\item{lobLengthThreshold}{is a length threshold (i.e. CL in centimeters) beyond which there is no chance of catching another lobster}

\item{lobSize}{is a size frequency dataset which is set to NA by default}

\item{sexBased}{is a logical parameter(TRUE/FALSE) which determines if sex of catch needs too be considered}

\item{lobSex}{is the sex of the lobster}

\item{r}{is the instantaneous rate of change in qo with respect to Ct}
}
\value{
Returns the probability of entry to trap
}
\description{
This function calculates the probability of entry to trap (aka catchability). In addition to parameters in Addison & Bell
1997, length of catch can is added and can be included in calculation of catchability.
}
