% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directionalMove.R
\name{directionalMove}
\alias{directionalMove}
\title{This function moves  lobsters toward the trap (+some randomness). The closer a lobster gets to the trap the smaller the random component of movement. This function will
a lobster is within the radius of influece of a trap.}
\usage{
directionalMove(
  Lobster,
  dStep,
  minDistoTrap,
  Trap,
  radiusOfInfluence,
  currentZoI
)
}
\arguments{
\item{Lobster}{location of lobster in x and y coordinates}

\item{dStep}{is how much a lobster moves in each time step}

\item{minDistoTrap}{is the distance to trap}

\item{Trap}{location of trap in x and y coordinates}

\item{radiusOfInfluence}{trap's radius of influence}

\item{currentZoI}{initial zone of influence}
}
\value{
Returns the new coordinates of each lobster
}
\description{
This function moves  lobsters toward the trap (+some randomness). The closer a lobster gets to the trap the smaller the random component of movement. This function will
a lobster is within the radius of influece of a trap.
}
