% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion.R
\name{dispersion}
\alias{dispersion}
\title{This function calculates the variance to mean ratio (AKA dispersion index)}
\usage{
dispersion(x)
}
\arguments{
\item{x}{is a numeric vector}
}
\value{
Returns the dispersion index
}
\description{
This function calculates the variance to mean ratio (AKA dispersion index)
}
