% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceToClosestTrap.R
\name{distanceToClosestTrap}
\alias{distanceToClosestTrap}
\title{This function returns id of the closest trap and calculates its distance to ? (Note: Complete later)}
\usage{
distanceToClosestTrap(Lobster, Trap)
}
\arguments{
\item{Lobster}{location of lobster in x and y coordinates}

\item{Trap}{location of trap in x and y coordinates}
}
\value{
Returns distance to closest trap and the trap ID
}
\description{
This function returns id of the closest trap and calculates its distance to ? (Note: Complete later)
}
