% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSimOutput.R
\name{GetSimOutput}
\alias{GetSimOutput}
\title{This function extracts two pieces of information at the end of simulation: maximum numbr of trapped
lobster and the time when that occurred.}
\usage{
GetSimOutput(x)
}
\arguments{
\item{x}{is an object that contains information from function SimulateLobsterMovement}
}
\value{
Returns  maximum catch and time to maximum catch
}
\description{
This function extracts two pieces of information at the end of simulation: maximum numbr of trapped
lobster and the time when that occurred.
}
