% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialLobsterGrid.R
\name{initialLobsterGrid}
\alias{initialLobsterGrid}
\title{This function simulates an arena and a number of lobsters to start with.}
\usage{
initialLobsterGrid(nrowgrids, ncolgrids, initlambda, initD, losbterSizeFile)
}
\arguments{
\item{nrowgrids}{is a numeric value which defines number of rows of the arena}

\item{ncolgrids}{is a numeric value which defines number of columns of the arena}

\item{initlambda}{is the density of lobsters}

\item{initD}{is the dispersion index of lobsters}

\item{losbterSizeFile}{is a csv file that contains the size frequency of lobsters}
}
\value{
Returns  x and y coordinates of simulated lobsters at the beginning
}
\description{
This function simulates an arena and a number of lobsters to start with.
}
