% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomMove.R
\name{randomMove}
\alias{randomMove}
\title{The function randomly selects an angle (0:360) and moves the lobster. This is called when a lobster
is outside the area of influence.}
\usage{
randomMove(Lobster, dStep)
}
\arguments{
\item{Lobster}{location of lobster in x and y coordinates}

\item{dStep}{is how much a lobster moves in each time step}
}
\value{
Returns the new coordinates of each lobster
}
\description{
The function randomly selects an angle (0:360) and moves the lobster. This is called when a lobster
is outside the area of influence.
}
