% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicateCoordinates.R
\name{replicateCoordinates}
\alias{replicateCoordinates}
\title{This function replicates the coordinates where there are multiple lobsters}
\usage{
replicateCoordinates(d)
}
\arguments{
\item{d}{is a data frame containing  x and y coordinates of lobsters and number of lobsters at each coordinate}
}
\value{
Returns a data frame
}
\description{
This function replicates the coordinates where there are multiple lobsters
}
