% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateLobsterMovement.R
\name{SimulateLobsterMovement}
\alias{SimulateLobsterMovement}
\title{A function to run the simulation based on defined parameters}
\usage{
SimulateLobsterMovement(p)
}
\arguments{
\item{p}{is a list which contains all input variables}
}
\value{
Returns a list
}
\description{
A function to run the simulation based on defined parameters
}
