% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapInPath.R
\name{trapInPath}
\alias{trapInPath}
\title{This function determines if lobster passes through a trap and determines if it is caught}
\usage{
trapInPath(loc1, loc2, Trap, howClose)
}
\arguments{
\item{loc1}{is the location of lobster at the start of each time step}

\item{loc2}{is the location of lobster at the end of each time step}

\item{Trap}{is the location of trap}

\item{howClose}{sets the distance from the trap within which catch occurs}
}
\value{
Returns a vector that contain lobster path and if it is trapped or not
}
\description{
This function determines if lobster passes through a trap and determines if it is caught
}
