% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateGrid.R
\name{updateGrid}
\alias{updateGrid}
\title{This function updates the coordinate of each lobster at each timestep,
The distanceToTrap function calculates the distance to the closest trap}
\usage{
updateGrid(
  Lobster,
  Trap,
  trapCatch,
  lobSize,
  lobSex,
  radiusOfInfluence,
  currentZoI,
  dStep,
  howClose,
  q0,
  qmin,
  saturationThreshold,
  trapSaturation,
  lengthBased,
  lobLengthThreshold,
  sexBased
)
}
\arguments{
\item{Lobster}{is the x & y coordinates of each lobster}

\item{Trap}{is the x & y coordinates of trap(s)}

\item{trapCatch}{number of captured lobster}

\item{lobSize}{size of captured lobster}

\item{radiusOfInfluence}{is the initial radius of influence (I0 in the paper)}

\item{currentZoI}{is the bait's area of influence at each timestep}

\item{dStep}{is how much a lobster moves in each time step}

\item{howClose}{sets the distance from the trap within which catch occurs}

\item{q0}{is the initial probability of entry into an empty trap(i.e. 0.5)}

\item{qmin}{is the asymptotic minimum probability of entry (i.e. 0)}

\item{saturationThreshold}{is the number of lobsters in a trap at which the probability of
another lobster entering the trap drops to qmin}

\item{trapSaturation}{is a logical parameter}

\item{lengthBased}{is a logical parameter}

\item{lobLengthThreshold}{is a length threshold (i.e. CL in centimeters) beyond which there is no chance of catching another lobster}

\item{sexBased}{is a logical parameter}
}
\value{
a list of new coordinates, number of catch and their sizes
}
\description{
This function updates the coordinate of each lobster at each timestep,
The distanceToTrap function calculates the distance to the closest trap
}
