% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSimOutput.R
\name{GetSimOutput}
\alias{GetSimOutput}
\title{This function extracts two pieces of information at the end of simulation: maximum number of trapped
lobster and the time when that occurred.}
\usage{
GetSimOutput(x, mls = 82.5)
}
\arguments{
\item{x}{is an object that contains information from function SimulateLobsterMovement}

\item{mls}{is the minimum legal size for calculating catch weight}
}
\value{
Returns  maximum catch in number total weight, and legal wt and length of time to reach maximum catch
}
\description{
This function extracts two pieces of information at the end of simulation: maximum number of trapped
lobster and the time when that occurred.
}
